# apps/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Wayfii API
  version: 0.2.0
  description: Catálogo turístico, planificación y gamificación de Wayfii.
servers:
  - url: /
    description: Mismo origen que el cliente
security: []
...
paths:
...
  /v1/catalog/places:
    get:
      summary: Buscar lugares publicados
      tags: [Catalog]
      operationId: searchPlaces
      parameters:
        - {$ref: "#/components/parameters/City"}
        - {$ref: "#/components/parameters/Category"}
        - {$ref: "#/components/parameters/Latitude"}
        - {$ref: "#/components/parameters/Longitude"}
        - {$ref: "#/components/parameters/Radius"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: Lugares publicados
          content:
            application/json:
              schema:
                type: object
                required: [items, count]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/Place"}
                  count: {type: integer}
        "400": {$ref: "#/components/responses/BadRequest"}
    post:
      summary: Crear o actualizar un lugar
      tags: [Catalog]
      operationId: upsertPlace
      security: [{AdminBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Place"}
      responses:
        "201":
          description: Lugar creado o actualizado
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Place"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/catalog/events:
    get:
      summary: Buscar eventos publicados
      tags: [Catalog]
      operationId: searchEvents
      parameters:
        - {$ref: "#/components/parameters/City"}
        - {$ref: "#/components/parameters/Category"}
        - {$ref: "#/components/parameters/Latitude"}
        - {$ref: "#/components/parameters/Longitude"}
        - {$ref: "#/components/parameters/Radius"}
        - {$ref: "#/components/parameters/Limit"}
        - name: from
          in: query
          schema: {type: string, format: date-time}
        - name: to
          in: query
          schema: {type: string, format: date-time}
      responses:
        "200":
          description: Eventos publicados
          content:
            application/json:
              schema:
                type: object
                required: [items, count]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/Event"}
                  count: {type: integer}
        "400": {$ref: "#/components/responses/BadRequest"}
    post:
      summary: Crear o actualizar un evento
      tags: [Catalog]
      operationId: upsertEvent
      security: [{AdminBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Event"}
      responses:
        "201":
          description: Evento creado o actualizado
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Event"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/catalog/import/osm:
    post:
      summary: Importar lugares desde OpenStreetMap
      tags: [Catalog]
      operationId: importOpenStreetMapPlaces
      security: [{AdminBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination, center, radiusMeters]
              properties:
                destination: {type: string}
                center: {$ref: "#/components/schemas/GeoPoint"}
                radiusMeters:
                  type: integer
                  minimum: 100
                  maximum: 20000
      responses:
        "201":
          description: Resultado de importación
          content:
            application/json:
              schema:
                type: object
                required: [imported, places, source]
                properties:
                  imported: {type: integer}
                  places:
                    type: array
                    items: {$ref: "#/components/schemas/Place"}
                  source: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "502":
          description: Proveedor OSM no disponible
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /v1/workers/catalog/jobs:
    post:
      summary: Encolar una sincronización del catálogo
      description: |
        Inicia una ejecución asíncrona y secuencial contra las fuentes
        configuradas. Si ya existe un job activo equivalente, devuelve ese
        mismo job y agrega `X-Wayfii-Job-Reused: true`.
      tags: [Workers]
      operationId: enqueueCatalogSync
      security: [{AdminBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CatalogSyncRequest"}
            example:
              destination: Buenos Aires
              center:
                latitude: -34.6037
                longitude: -58.3816
              countryCode: AR
              radiusMeters: 5000
              sources: [openstreetmap, ticketmaster]
              eventWindowDays: 30
      responses:
        "202":
          description: Job aceptado o ejecución activa equivalente reutilizada
          headers:
            Location:
              description: Recurso para consultar el estado del job.
              schema: {type: string}
            X-Wayfii-Job-Reused:
              description: Presente y verdadero cuando no se creó un job nuevo.
              schema: {type: boolean}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CatalogSyncJob"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429":
          description: La cola acotada del worker está completa
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "503":
          description: Worker no disponible
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /v1/workers/catalog/jobs/{jobID}:
    get:
      summary: Consultar una sincronización del catálogo
      tags: [Workers]
      operationId: getCatalogSyncJob
      security: [{AdminBearer: []}]
      parameters:
        - name: jobID
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Estado actual o resultado final del job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CatalogSyncJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404":
          description: Job inexistente o removido del historial acotado
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "503":
          description: Worker no disponible
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /v1/itineraries/plan:
    post:
      summary: Planificar un itinerario
      tags: [Itineraries]
      operationId: planItinerary
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PlanItineraryRequest"}
      responses:
        "200":
          description: Itinerario planificado
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PlannedItinerary"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404":
          description: No hay lugares publicados para planificar
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /v1/gamification/activities:
    post:
      summary: Registrar una actividad idempotente
      tags: [Gamification]
      operationId: recordActivity
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Activity"}
      responses:
        "200":
          description: La clave de idempotencia ya había sido procesada
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RecordActivityResult"}
        "201":
          description: Actividad registrada
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RecordActivityResult"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /v1/gamification/players/{userID}:
    get:
      summary: Consultar progreso del viajero
      tags: [Gamification]
      operationId: getPlayer
      parameters:
        - name: userID
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Perfil gamificado
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PlayerProfile"}
        "400": {$ref: "#/components/responses/BadRequest"}
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
  parameters:
    City:
      name: city
      in: query
      schema: {type: string}
    Category:
      name: category
      in: query
      schema: {$ref: "#/components/schemas/Category"}
    Latitude:
      name: lat
      in: query
      schema: {type: number, minimum: -90, maximum: 90}
    Longitude:
      name: lon
      in: query
      schema: {type: number, minimum: -180, maximum: 180}
    Radius:
      name: radiusMeters
      in: query
      schema: {type: number, minimum: 1, maximum: 100000, default: 5000}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: Credenciales administrativas inválidas
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
    GeoPoint:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
    Category:
      type: string
      enum:
        - classic
        - alternative
        - popular
        - culture
        - gastronomy
        - nature
        - family
        - history
        - shopping
        - photography
        - events
        - adventure
    PriceLevel:
      type: string
      enum: [free, low, medium, high]
    PublicationStatus:
      type: string
      enum: [draft, published, archived]
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: {type: number, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3}
    Source:
      type: object
      required: [provider]
      properties:
        provider: {type: string}
        externalId: {type: string}
        license: {type: string}
        attribution: {type: string}
        sourceUrl: {type: string, format: uri}
    Place:
      type: object
      required:
        - name
        - summary
        - location
        - city
        - categories
        - priceLevel
        - recommendedDurationMinutes
        - qualityScore
        - source
      properties:
        id: {type: string}
        slug: {type: string}
        name: {type: string}
        summary: {type: string}
        description: {type: string}
        location: {$ref: "#/components/schemas/GeoPoint"}
        address: {type: string}
        city: {type: string}
        countryCode: {type: string, minLength: 2, maxLength: 2}
        categories:
          type: array
          minItems: 1
          items: {$ref: "#/components/schemas/Category"}
        tags:
          type: object
          additionalProperties: {type: string}
        priceLevel: {$ref: "#/components/schemas/PriceLevel"}
        estimatedCostPerPerson: {$ref: "#/components/schemas/Money"}
        recommendedDurationMinutes: {type: integer, minimum: 0}
        qualityScore: {type: number, minimum: 0, maximum: 1}
        imageUrl: {type: string, format: uri}
        status: {$ref: "#/components/schemas/PublicationStatus"}
        source: {$ref: "#/components/schemas/Source"}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
    Event:
      type: object
      required:
        - name
        - location
        - city
        - categories
        - startsAt
        - endsAt
        - priceLevel
        - source
      properties:
        id: {type: string}
        placeId: {type: string}
        name: {type: string}
        description: {type: string}
        location: {$ref: "#/components/schemas/GeoPoint"}
        address: {type: string}
        city: {type: string}
        countryCode: {type: string, minLength: 2, maxLength: 2}
        categories:
          type: array
          items: {$ref: "#/components/schemas/Category"}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        priceLevel: {$ref: "#/components/schemas/PriceLevel"}
        ticketUrl: {type: string, format: uri}
        imageUrl: {type: string, format: uri}
        status: {$ref: "#/components/schemas/PublicationStatus"}
        source: {$ref: "#/components/schemas/Source"}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
    CatalogSource:
      type: string
      enum: [openstreetmap, ticketmaster]
    CatalogSyncStatus:
      type: string
      enum: [queued, running, succeeded, partial, failed, cancelled]
    CatalogSyncTrigger:
      type: string
      enum: [api, config]
    CatalogSyncRequest:
      type: object
      required: [destination]
      properties:
        destination:
          type: string
          minLength: 1
          description: Nombre normalizado de la ciudad o destino.
        center:
          allOf:
            - {$ref: "#/components/schemas/GeoPoint"}
          description: |
            Centro de búsqueda. Si se omite en una llamada manual, Wayfii usa
            Nominatim. Es obligatorio para ejecuciones configuradas.
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
          description: Código ISO 3166-1 alpha-2 usado para acotar eventos.
        radiusMeters:
          type: integer
          minimum: 100
          maximum: 20000
          default: 5000
        sources:
          type: array
          uniqueItems: true
          description: |
            Fuentes a consultar. Si se omite, se usan todas las fuentes
            configuradas en el runtime.
          items: {$ref: "#/components/schemas/CatalogSource"}
        eventWindowDays:
          type: integer
          minimum: 1
          maximum: 90
          default: 30
...
    CatalogSyncJob:
      type: object
      required: [id, status, trigger, request, results, createdAt]
      properties:
        id: {type: string}
        status: {$ref: "#/components/schemas/CatalogSyncStatus"}
        trigger: {$ref: "#/components/schemas/CatalogSyncTrigger"}
        request: {$ref: "#/components/schemas/CatalogSyncRequest"}
        results:
          type: array
          items: {$ref: "#/components/schemas/CatalogSourceSyncResult"}
        error: {type: string}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time}
    PlanItineraryRequest:
      type: object
      required:
        - destination
        - interests
        - mobility
        - startMinutes
        - endMinutes
        - people
        - budget
      properties:
        destination: {type: string}
        center: {$ref: "#/components/schemas/GeoPoint"}
        origin: {$ref: "#/components/schemas/GeoPoint"}
        originName:
          type: string
          description: Dirección o zona que la API geocodificará si origin no está presente.
        interests:
          type: array
          minItems: 1
          items: {$ref: "#/components/schemas/Category"}
        mobility:
          type: array
          minItems: 1
          items: {type: string}
        startMinutes: {type: integer, minimum: 0, maximum: 1439}
        endMinutes: {type: integer, minimum: 1, maximum: 1440}
        people: {type: integer, minimum: 1, maximum: 20}
        budget: {$ref: "#/components/schemas/PriceLevel"}
        pace:
          type: string
          enum: [relaxed, balanced, intense]
          default: balanced
        includeFood: {type: boolean, default: true}
...
    PlannedItinerary:
      type: object
      required:
        - destination
        - center
        - origin
        - stops
        - startMinutes
        - endMinutes
        - dataSourceSummary
      properties:
        destination: {type: string}
        center: {$ref: "#/components/schemas/GeoPoint"}
        origin: {$ref: "#/components/schemas/GeoPoint"}
        stops:
          type: array
          items: {$ref: "#/components/schemas/ItineraryStop"}
        startMinutes: {type: integer}
        endMinutes: {type: integer}
        estimatedCost: {$ref: "#/components/schemas/Money"}
        dataSourceSummary: {type: string}
    Activity:
      type: object
      required: [idempotencyKey, userId, type, subjectId, occurredAt]
      properties:
        idempotencyKey:
          type: string
          maxLength: 160
          description: Clave estable y única por usuario, sujeto y acción.
        userId:
          type: string
          maxLength: 160
          description: Identificador pseudónimo del viajero; será reemplazado por el sujeto autenticado.
        type:
          type: string
          enum:
            - place_visited
            - event_attended
            - itinerary_completed
            - place_contributed
            - place_validated
            - offer_redeemed
        subjectId:
          type: string
          maxLength: 160
          description: Lugar, evento, itinerario, contribución u oferta asociada.
        metadata:
          type: object
          maxProperties: 16
          additionalProperties: {type: string}
        occurredAt: {type: string, format: date-time}
    PlayerProfile:
      type: object
      required: [userId, points, level, currentStreak, badges, updatedAt]
      properties:
        userId: {type: string}
        points: {type: integer}
        level: {type: integer}
        currentStreak: {type: integer}
        badges:
          type: array
          items: {type: string}
        lastActivityAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    RecordActivityResult:
      type: object
      required: [recorded, awardedPoints, earnedBadges, profile]
      properties:
        recorded: {type: boolean}
        awardedPoints:
          type: integer
          minimum: 0
          description: Delta otorgado por esta escritura; es cero en reintentos idempotentes.
        earnedBadges:
          type: array
          description: Badges desbloqueados por esta actividad.
          items: {type: string}
        profile: {$ref: "#/components/schemas/PlayerProfile"}